requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart==0.0.9
numpy>=1.24.0
pandas>=2.0.0
requests==2.31.0
python-jose>=3.3.0
python-dotenv>=1.0.0
albumentations>=1.3.1
matplotlib>=3.7.0
seaborn>=0.12.0
aiofiles>=23.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1
passlib>=1.7.4
streamlit==1.31.0
pillow==10.2.0
